refactor : (로또 게임) 보너스 번호가 해당 되는 로또 코드 리팩토링
