+++ src/main/java/lotto/domain/LottoGameMachine.java
@@ -1,6 +1,7 @@
 package lotto.domain;
 
 import lotto.Lotto;
+import lotto.domain.subaction.WinningNumberPlace;
 import lotto.input.AdditionalNumber;
 import lotto.input.LottoTickets;
 import lotto.input.WinningNumber;
@@ -24,12 +25,11 @@ public class LottoGameMachine {
         List<List<Integer>> randomLotto = randomLottoGenerator.createTicket(lottos); // 랜덤 받기
         Lotto lottoAnswer = winningNumber.getCorrect(); // 정답 받기
 
-        LottoAndWinningNumber lottoAndWinningNumber = new LottoAndWinningNumber();
-        List<Integer> sameNumberCount = lottoAndWinningNumber.comparison(randomLotto, lottoAnswer); // 결과 받기
+        WinningNumberAndLotto winningNumberAndLotto = new WinningNumberAndLotto(new WinningNumberPlace());
+        List<Integer> sameNumberCount = winningNumberAndLotto.comparison(randomLotto, lottoAnswer); // 결과 받기
 
         LottoGameManagement lottoGameManagement = new LottoGameManagement(new AdditionalNumber());
         lottoGameManagement.bonusConfirmation(randomLotto, lottoAnswer, sameNumberCount);
-
 
         //return 0;
     }
